fix: add browser support
