+++ jest.config.js
@@ -4,6 +4,6 @@ module.exports = {
   resetMocks: true,
   restoreMocks: true,
   rootDir: './src',
-  testEnvironment: 'node',
-  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  preset: 'ts-jest'
 };
