+++ package.json
@@ -4,6 +4,7 @@
   "version": "0.0.0-semantic-release-dev",
   "description": "",
   "main": "build/src/index.js",
+  "browser": "build/src/index-web.js",
   "scripts": {
     "build": "npm run build:code && npm run build:docs",
     "build:docs": "typedoc",
@@ -29,16 +30,16 @@
   },
   "homepage": "https://github.com/open-rpc/schema-utils-js#readme",
   "dependencies": {
-    "@open-rpc/meta-schema": "^1.3.1",
+    "@open-rpc/meta-schema": "^1.3.2",
     "ajv": "^6.10.0",
     "detect-node": "^2.0.4",
     "fs-extra": "^7.0.1",
     "is-url": "^1.2.4",
-    "json-schema-ref-parser": "^6.0.3",
-    "node-fetch": "^2.3.0"
+    "isomorphic-fetch": "^2.2.1",
+    "json-schema-ref-parser": "^6.0.3"
   },
   "devDependencies": {
-    "@qiwi/semantic-release-gh-pages-plugin": "^1.6.2",
+    "@qiwi/semantic-release-gh-pages-plugin": "^1.9.1",
     "@semantic-release/changelog": "^3.0.2",
     "@semantic-release/commit-analyzer": "^6.1.0",
     "@semantic-release/git": "^7.0.8",
@@ -48,6 +49,7 @@
     "@types/detect-node": "^2.0.0",
     "@types/fs-extra": "^7.0.0",
     "@types/is-url": "^1.2.28",
+    "@types/isomorphic-fetch": "0.0.35",
     "@types/jest": "^24.0.11",
     "@types/json-schema": "^7.0.3",
     "@types/lodash": "^4.14.123",
