+++ src/get-open-rpc-document-from-url.ts
@@ -1,0 +1,14 @@
+import { OpenRPC } from "@open-rpc/meta-schema";
+import fetch from "isomorphic-fetch";
+import { TGetOpenRPCDocument } from "./get-open-rpc-document";
+
+const fetchUrlSchema: TGetOpenRPCDocument = async (schema) => {
+  try {
+    const response = await fetch(schema);
+    return await response.json() as OpenRPC;
+  } catch (e) {
+    throw new Error(`Unable to download openrpc.json file located at the url: ${schema}`);
+  }
+};
+
+export default fetchUrlSchema;
