+++ tsconfig.json
@@ -3,7 +3,8 @@
     "target": "es5",
     "module": "commonjs",
     "lib": [
-      "ES2015"
+      "es2015",
+      "dom"
     ],
     "declaration": true,
     "outDir": "./build",
